+++ src/vuex/store.js
@@ -7,10 +7,12 @@ export default new Vuex.Store({
   /**
    * Vuex state
    * @alias VuexState
+   * @property {string} loaderIconFilter - defines svg-filter with color settings for loader icon
    * @property {string|null} searchPattern - pattern of inner search for available options
    * @property {string|null} searchValue - value of search field
    */
   state: {
+    loaderIconFilter: "loader-default",
     searchPattern: null,
     searchValue: null,
   },
@@ -20,6 +22,16 @@ export default new Vuex.Store({
    * @alias VuexGetters
    */
   getters: {
+    /**
+     * Getter for loaderIconFilter property
+     * @method
+     * @param {Object} state - {@linkcode VuexState}
+     * @returns {string} filter
+     */
+    loaderIconFilter(state) {
+      return state.loaderIconFilter;
+    },
+
     /**
      * Getter for searchPattern property
      * @method
@@ -47,6 +59,16 @@ export default new Vuex.Store({
    */
   mutations: {
     /**
+     * Mutation of loaderIconFilter property
+     * @method SET_LOADER_ICON_FILTER
+     * @param {Object} state - {@linkcode VuexState}
+     * @param {string} pfilter - new svg-filter with color settings for loader icon
+     */
+    SET_LOADER_ICON_FILTER(state, filter) {
+      state.loaderIconFilter = filter;
+    },
+
+    /**
      * Mutation of searchPattern property
      * @method SET_SEARCH_PATTERN
      * @param {Object} state - {@linkcode VuexState}
@@ -72,6 +94,17 @@ export default new Vuex.Store({
    * @alias VuexActions
    */
   actions: {
+    /**
+     * Action triggers {@linkcode SET_LOADER_ICON_FILTER} mutation
+     * @method
+     * @param {Object} store - Vuex store API
+     * @param {Function} store.commit - method for mutation triggering
+     * @param {string} pattern - new svg-filter with color settings for loader icon
+     */
+    setLoaderIconFilter({ commit }, filter) {
+      commit("SET_LOADER_ICON_FILTER", filter);
+    },
+
     /**
      * Action triggers {@linkcode SET_SEARCH_PATTERN} mutation
      * @method